src/requirements.txt
torch
torchvision 
numpy
opencv-python
scikit-image
matplotlib
seaborn
tqdm
Pillow
transformers